feat: support `build` command for static hosting
